Futex - don't check atomics for WASI
We don't do the check to see if we can use atomics inside futex anymore. This is passed as an argument from a futex_wrapper instead.
