+++ system/common.cpp
@@ -156,7 +156,7 @@ long WEAK __syscall_futex(uint32_t* uaddr, int futex_op, ...)
 {
 	va_list args;
 	va_start(args, futex_op);
-	int ret = sys_internal::futex(uaddr, futex_op, args);
+	int ret = sys_internal::futex_wrapper(uaddr, futex_op, args);
 	va_end(args);
 	return ret;
 }
@@ -423,7 +423,7 @@ long WEAK sched_getaffinity(pid_t pid, int cpusetsize, unsigned long* mask)
 	return cpusetsize;
 }
 
-long WEAK futex(uint32_t* uaddr, int futex_op, va_list args)
+long WEAK futex(uint32_t* uaddr, int futex_op, bool canUseAtomics, va_list args)
 {
 	return -ENOSYS;
 }
