+++ system/impl.h
@@ -15,8 +15,10 @@ namespace sys_internal {
 	long tkill(pid_t tid, int sig);
 	long exit_group(long code);
 	long sched_getaffinity(pid_t pid, int cpusetsize, unsigned long* mask);
-	long futex(uint32_t* uaddr, int futex_op, va_list args);
+	long futex_wrapper(uint32_t* uaddr, int futex_op, va_list args);
+	long futex(uint32_t* uaddr, int futex_op, bool canUseAtomics, va_list args);
 	[[cheerp::wasm]] void killAllThreads();
+	bool isBrowserMainThread();
 }
 
 extern _Thread_local int tid;
